+++ src/main/java/de/schaack/ml/basics/functions/activation/interfaces/ActivationFunction.java
@@ -3,5 +3,7 @@ package de.schaack.ml.basics.functions.activation.interfaces;
 public interface ActivationFunction {
    double activate(double value);
 
-   double deriveActivation(double value);
+   double deriveActivation(double globalDerivative);
+
+   double getCurrentGradient();
 }