+++ src/main/java/de/schaack/ml/basics/functions/loss/implementation/BinaryCrossEntropyLoss.java
@@ -11,15 +11,18 @@ public class BinaryCrossEntropyLoss implements LossFunction {
 
     private static final Logger log = LoggerFactory.getLogger(BinaryCrossEntropyLoss.class);
 
+    private double currentValue = 0;
+    private double currentGradient = 0;
+
     @Override
     public double calculateLoss(double trueLabel, double predictedLabel) throws IllegalArgumentException {
         if (predictedLabel <= 0.0 || predictedLabel >= 1.0)
             throw new IllegalArgumentException("PredictedLabel must be within 0 and 1: " + predictedLabel);
 
-        double loss = -(trueLabel * Math.log(predictedLabel)) - (1 - trueLabel * Math.log(1 - predictedLabel));
+        currentValue = -(trueLabel * Math.log(predictedLabel)) - (1 - trueLabel * Math.log(1 - predictedLabel));
         log.debug("The loss of trueLabel {} with predictedLabel {} is calculated as {}", trueLabel, predictedLabel,
-                loss);
-        return loss;
+                currentValue);
+        return currentValue;
     }
 
     @Override
@@ -27,10 +30,10 @@ public class BinaryCrossEntropyLoss implements LossFunction {
         if (predictedLabel <= 0.0 || predictedLabel >= 1.0)
             throw new IllegalArgumentException("PredictedLabel must be within 0 and 1: " + predictedLabel);
 
-        double derivedValue = -((trueLabel / predictedLabel) - ((1 - trueLabel) / (1 - predictedLabel)));
+        currentGradient = -((trueLabel / predictedLabel) - ((1 - trueLabel) / (1 - predictedLabel)));
         log.debug("The derived loss of trueLabel {} with predictedLabel {} is calculated as {}", trueLabel,
-                predictedLabel, derivedValue);
-        return derivedValue;
+                predictedLabel, currentGradient);
+        return currentGradient;
     }
 
     @Override