+++ src/main/java/de/schaack/ml/basics/models/implementation/classification/Perceptron.java
@@ -36,6 +36,7 @@ public class Perceptron implements Model<Double> {
         if (getParameters().length != dataPoint.getEntries().length)
             throw new IllegalArgumentException("Weights and dataPoint are not of equal size.");
 
+            //TODO remove bias from datapoint
         this.currentSumOfProducts = MatrixOperations.valuesMulData(getParameters(), dataPoint);
         double currentActivatedSumOfProducts = perceptronModelSettings.getActivationFunction()
                 .activate(currentSumOfProducts);
@@ -48,16 +49,18 @@ public class Perceptron implements Model<Double> {
     // TODO Adjust for variably sized outputOfPreviousComponent (Layer-to-Layer)
     @Override
     public double[] backPropagate(DataPoint currentDataPoint, double outputOfPreviousComponent) {
-        double activatedSumOfProductsDerivative = getModelSettings()
+        getModelSettings()
                 .getActivationFunction()
-                .deriveActivation(this.currentSumOfProducts);
+                .deriveActivation(outputOfPreviousComponent);
         double[] parameterDerivatives = currentDataPoint.getEntries();
-        this.currentParameterGradients = MatrixOperations
-                .valuesMulValue(parameterDerivatives, activatedSumOfProductsDerivative * outputOfPreviousComponent);
+        this.currentParameterGradients = 
+                        MatrixOperations.valuesMulValue(parameterDerivatives,
+                                getModelSettings().getActivationFunction().getCurrentGradient());
 
         log.debug(
                 "By multiplying and summing parameterDerivatives {} with (datapoint {} * previousGradient {})  calculates currentParameterGradients as {}",
-                parameterDerivatives, currentDataPoint.getEntries(), outputOfPreviousComponent,
+                parameterDerivatives, currentDataPoint.getEntries(),
+                getModelSettings().getActivationFunction().getCurrentGradient(),
                 this.currentParameterGradients);
 
         return this.currentParameterGradients;
@@ -81,11 +84,9 @@ public class Perceptron implements Model<Double> {
      */
     @Override
     public Double predict(DataPoint dataPoint) {
-        if (dataPoint.hasLabel())
-            throw new IllegalArgumentException("DataPoint has to be unlabeled.");
-        else if (getParameters().length != dataPoint.getEntries().length)
+        if (getParameters().length != dataPoint.getEntries().length)
             throw new IllegalArgumentException("Weights and dataPoint are not of equal size.");
-        
+
         this.getModelSettings().getDataPreprocessor().preprocessDatapoint(dataPoint);
 
         double matMulResult = MatrixOperations.valuesMulData(getParameters(), dataPoint);
@@ -101,7 +102,7 @@ public class Perceptron implements Model<Double> {
     }
 
     @Override
-    public <S extends DataSet<? extends DataPoint>>  Double[] predictions(S dataSet) {
+    public <S extends DataSet<? extends DataPoint>> Double[] predictions(S dataSet) {
         return predictions(dataSet.getDataPoints());
     }